https://github.com/osexpert/ccs32clara-chademo/issues/33#issuecomment-3436009872
ZE0 set CarStatus::UNKNOWN_102_5_6. Can this possibly mean that it support dynamic control even without  110.ExtendedFunction1? Try it...

Tried: ccs32clara-chademo 20251002-175:
There is only one example of it in the log, but it shows the car is reducing the requested current when AvailableOutputCurrent is lowered

From what I can see, this car do support dynamic AvailableOutputCurrent, even if it does not declare it (according to spec).
So I guess/assume the undocumented status flag the car set, is the car secretly telling it does support dynamic AvailableOutputCurrent after all.