https://site-assets.evgo.com/f/78437/x/e2ecd6d60d/evgo-vehicle-oem-best-practices.pdf
2.5. Inclusion of Maximum Power Limit
Vehicles shall include the maximum power rating of the vehicle in charge parameter discovery and
current demand.
In DIN 70121 (Table 71) and ISO 15118-2 (Table 101), maximum power limit is considered an
optional parameter. However, this parameter can be useful for the EVSE to provide accurate power
allocation. The parameter is mandatory in ISO 15118-20.